# docs/swagger.yaml
definitions:
  dto.CreatePaymentRequest:
    properties:
      amount:
        type: number
      currency:
        $ref: '#/definitions/dto.PaymentCurrency'
      reference:
        type: string
    type: object
  dto.CreatePaymentResponse:
    properties:
      id:
        type: string
      status:
        $ref: '#/definitions/dto.PaymentStatus'
    type: object
  dto.GetPaymentDetailsResponse:
    properties:
      amount:
        type: number
      created_at:
        type: string
      currency:
        $ref: '#/definitions/dto.PaymentCurrency'
      id:
        type: string
      reference:
        type: string
      status:
        $ref: '#/definitions/dto.PaymentStatus'
    type: object
  dto.PaymentCurrency:
    enum:
    - ETB
    - USD
    type: string
    x-enum-varnames:
    - ETB
    - USD
  dto.PaymentStatus:
    enum:
    - PENDING
    - SUCCESS
    - FAILED
    type: string
    x-enum-varnames:
    - PENDING
    - SUCCESS
    - FAILED
info:
  contact: {}
paths:
  /api/v1/payments:
    post:
      consumes:
      - application/json
      description: Creates a new payment record and initiates processing via RabbitMQ
      parameters:
      - description: Payment creation request
        in: body
        name: payment
        required: true
        schema:
          $ref: '#/definitions/dto.CreatePaymentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CreatePaymentResponse'
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new payment
      tags:
      - Payments
  /api/v1/payments/{id}:
    get:
      description: Retrieves details of a payment by its unique ID
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GetPaymentDetailsResponse'
        "400":
          description: Invalid ID format
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Payment not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get payment details
      tags:
      - Payments
swagger: "2.0"